+++ Engine/Mesh.h
@@ -11,12 +11,12 @@ public:
 		float3 position;
 		float2 tex_coords;
 	};
-	Mesh(std::vector<Vertex> && vertices, std::vector<UINT64> && indices);
-	Mesh(std::vector<Vertex> && vertices, std::vector<UINT64> && indices, std::string mesh_file_path);
+	Mesh(std::vector<Vertex> && vertices, std::vector<UINT32> && indices);
+	Mesh(std::vector<Vertex> && vertices, std::vector<UINT32> && indices, std::string mesh_file_path);
 	~Mesh();
 
 	std::vector<Vertex> vertices;
-	std::vector<UINT64> indices;
+	std::vector<UINT32> indices;
 	std::string mesh_file_path;
 	int num_triangles = 0;
 	int num_vertices = 0;
